Wayland display server
Introduce a new display server option to run the greeter with.
In the future this will be the default display server.

This patch contains the most important parts to get the greeter running
on Wayland, but it's not possible to select a keyboard layout.

At the moment the Wayland socket is assumed to be "wayland-0" but in a
multi-seat configuration we would have more than one greeter running,
each with a diffent socket and the assumption won't work anymore.

In the future we will figure out how to do, after all at the moment this
is considered enough for most people.

Closes: #440